+++ src/surface/gui/gui/node_singleton.py
