+++ src/surface/gui/gui/widgets/heartbeat.py
@@ -4,8 +4,7 @@ from PyQt6.QtWidgets import QHBoxLayout, QLabel, QVBoxLayout, QWidget
 from rov_msgs.msg import VehicleState
 
 from gui.widgets.circle import CircleIndicator
-from gui.widgets.node_singleton import GUINode
-
+from gui.node_singleton import GUINode
 
 class HeartbeatWidget(QWidget):
     signal = pyqtSignal(VehicleState)
