+++ src/surface/gui/gui/widgets/logger.py
@@ -4,8 +4,7 @@ from PyQt6.QtWidgets import QCheckBox, QHBoxLayout, QTextEdit, QVBoxLayout, QWid
 from rcl_interfaces.msg import Log
 from rclpy.impl.logging_severity import LoggingSeverity
 
-from gui.widgets.node_singleton import GUINode
-
+from gui.node_singleton import GUINode
 # Dictionary linking LoggingSeverity to a QColor
 SEVERITY_LEVELS_DICT = {
     LoggingSeverity.UNSET: QColor(0, 0, 0),
