+++ src/surface/gui/gui/widgets/timer.py
@@ -5,8 +5,7 @@ from rclpy.duration import Duration
 from rov_msgs.msg import MissionTimerTick
 from rov_msgs.srv import MissionTimerSet
 
-from gui.widgets.node_singleton import GUINode
-
+from gui.node_singleton import GUINode
 RESET_SECONDS = 15 * 60  # The number of seconds to set the timer to when reset is clicked
